cufile: use absolute includes, cuda::std traits; inline trivial defs; tighten capability checks Switch to <...> absolute includes across cufile headers/impls Prefer cuda::std::is_trivially_copyable_v over std:: Inline batch_io_params_span ctor and batch_io_result helpers in batch_handle.hpp; remove redundant impls Replace try/catch in capability checks with direct cuFile C API checks Minor: include <sys/types.h> for off_t; tidy buffer registration logic
