+++ cudax/include/cuda/experimental/__cufile/buffer_handle.hpp
@@ -23,8 +23,8 @@
 
 #include <functional>
 
-#include "detail/error_handling.hpp"
-#include "detail/raii_resource.hpp"
+#include <cuda/experimental/__cufile/detail/error_handling.hpp>
+#include <cuda/experimental/__cufile/detail/raii_resource.hpp>
 
 namespace cuda::experimental::cufile
 {
@@ -104,4 +104,4 @@ public:
 
 } // namespace cuda::experimental::cufile
 
-#include "detail/buffer_handle_impl.hpp"
+#include <cuda/experimental/__cufile/detail/buffer_handle_impl.hpp>
